requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy
pdfminer.six
PyPDF2
matplotlib
google-generativeai
python-docx
